Moved AqlExecuteResult to a separate file
